Split Landing Pages test and Locked Pages tests
